+++ benight-admin/src/app/core/interfaces/interfaces.database/interfaces.public-rules.ts
@@ -1,0 +1,6 @@
+//public rules por evento
+//public rules por evento fecha
+//public rules por plan
+//public rules por plan fecha
+//public rules por entrada
+//public rules por entrada fecha